+++ docs/changelog.rst
@@ -2,7 +2,8 @@
 Changelog
 =========
 
-* :feature:`` Users can now manually link assets on any counterparty to assets recognized by Rotki, without having to wait for a new release.
+* :feature:`-` GG23 arbitrum OSS rounds and impact donation minting will now be properly recognized by rotki
+* :feature:`-` Users can now manually link assets on any counterparty to assets recognized by rotki, without having to wait for a new release.
 * :feature:`-` rotki will show the oracle used to display the asset price.
 * :feature:`8928` Users will now have the option to keep the table sorting preference.
 * :feature:`9495` rotki now supports Stake DAO across all supported EVM chains, including balance tracking.